+++ Year 4/NG4D900 - MEng Group Project/Renesas-MCU-Car-Rally-2018-USW/Software/MiconCarRally/lib/drivers/sound/piezo.h
@@ -13,14 +13,14 @@
 #include <libs/spwm/spwm.h>
 
 typedef struct {
-	uint32_t pitch;
+	float    pitch;
 	uint32_t duration;
 } note_t;
 
 typedef struct {
 	spwm_t * dev_handle;
 	note_t   note;
-	uint8_t  volume;
+	float    volume;
 	bool     is_playing;
 	bool     stop_playing;
 } piezo_t;
@@ -51,7 +51,7 @@ enum PIEZO_RETCODE piezo_play(piezo_t * handle, note_t * note, bool serial);
 enum PIEZO_RETCODE piezo_play_song_async(piezo_t * handle, note_t * song, uint32_t sequence_count, bool serial);
 enum PIEZO_RETCODE piezo_play_song_async_backwards(piezo_t * handle, note_t * song, uint32_t sequence_count, bool serial);
 enum PIEZO_RETCODE piezo_stop(piezo_t * handle, bool stop_and_wait);
-enum PIEZO_RETCODE piezo_set_volume(piezo_t * handle, uint8_t volume);
+enum PIEZO_RETCODE piezo_set_volume(piezo_t * handle, float volume);
 
 enum MIDI_LUT      midi_lut(uint8_t tone_index);
 uint16_t           midi_lut_count(void);
